+++ packages/blogpress/.vitepress/config.ts
@@ -1,14 +1,12 @@
 import process from 'node:process'
 import { getThemeConfig } from '@sugarat/theme/node'
 import type { Theme } from '@sugarat/theme'
-import type { RSSOptions } from 'vitepress-plugin-rss'
-import { RssPlugin } from 'vitepress-plugin-rss'
 import { defineConfig } from 'vitepress'
 
 const baseUrl = process.env.PAGES_BASE ?? 'https://www.dmsrs.org'
 const relativeUrl = process.env.PUBLIC_URL ?? "/"
 const weekly = `${baseUrl}/weekly`
-const RSSWeekly: RSSOptions = {
+const RSSWeekly: Theme.RSSOptions = {
   title: '视野修炼 - 技术周刊',
   baseUrl,
   description: '每周会精选出一些 优质&有趣 的内容做推送（大前端为主），包含但不限于 优质文章，开源库，工具网站，有意思的知识',
@@ -126,7 +124,9 @@ const blogTheme = getThemeConfig({
     random: true,
     limit: 6,
   },
-  // search: false,
+  search: {
+    showDate: true,
+  },
   recommend: {
     showSelf: true,
     nextText: '下一页',
@@ -225,9 +225,6 @@ export default defineConfig({
       port: 4000,
       host: '0.0.0.0',
     },
-    plugins: [
-      RssPlugin(RSSWeekly),
-    ],
   },
   vue: {
     template: {